REST Implementation Before Unit Testing
REST implementation completed

I had a small issue converting the string date into a date object while retaining the required "MM/DD/YYYY". As lists are unsorted, and I decided to use a list of dictionaries, I understand I'd need to iterate through each item and create / display a sorted list by the date using the lambda function. I was having a bit of a syntax issue, so decided not to let that hang me up.

There's a bit more error handling that could be done on the REST portion, but given that it only takes in one record at a time, I went under the assumption the user will easily be able to review their input to the app.

Next steps are to finalize unit testing, and while I understand the logic, I am not a developer by trade so this will most certainly be a learn by doing process. However, worst case scenario I wanted to provide everything up to now. 

I also supplied a small text file to use for testing.