Unit Testing + Updates to Errors Found
Unit testing revealed some errors in the logic of checking the data being passed through. Some new try/catch blocks were added as well as some value errors for future unit testing.

Dates are now being sorted and displayed in the expected fashion. First we convert the date string to a date object... then right back to a string. This is simply to make sure it's in the right format. Then when displaying the sorted list by date, we sort again by converting it to a date object. This is solely to prevent creating a new list or modifying the data in the list.

Still have some figuring out to do on unit testing for "ingest.py" which takes in a file, as well as the REST portion. Hoping to update some more unit testing prior to interview to get closer to the 80% coverage.

And finally, made the outputted JSON a bit more readable by adding indents on the outputs.